Refactor: getMailList - Query package 로 이동
