+++ src/main/java/ticketaka/mtvs3_final_backend/mail/query/controller/MailQueryController.java
@@ -1,5 +1,6 @@
-package ticketaka.mtvs3_final_backend.mail.command.application.controller;
+package ticketaka.mtvs3_final_backend.mail.query.controller;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -8,7 +9,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ticketaka.mtvs3_final_backend._core.utils.ApiUtils;
 import ticketaka.mtvs3_final_backend.mail.command.application.dto.MailCommandResponseDTO;
-import ticketaka.mtvs3_final_backend.mail.command.application.service.MailCommandService;
+import ticketaka.mtvs3_final_backend.mail.query.dto.MailQueryResponseDTO;
+import ticketaka.mtvs3_final_backend.mail.query.service.MailQueryService;
 
 import static ticketaka.mtvs3_final_backend._core.utils.SecurityUtils.getCurrentMemberId;
 
@@ -16,9 +18,9 @@ import static ticketaka.mtvs3_final_backend._core.utils.SecurityUtils.getCurrent
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/member/mails")
-public class MailCommandController {
+public class MailQueryController {
 
-    private final MailCommandService mailCommandService;
+    private final MailQueryService mailQueryService;
 
     /*
         우편 리스트 조회
@@ -26,7 +28,7 @@ public class MailCommandController {
     @GetMapping
     public ResponseEntity<?> getMailList() {
 
-        MailCommandResponseDTO. responseDTO = mailCommandService.(getCurrentMemberId());
+        MailQueryResponseDTO.getMailListDTO responseDTO = mailQueryService.getMailList(getCurrentMemberId());
 
         return ResponseEntity.ok().body(ApiUtils.success(responseDTO));
     }
