+++ src/main/java/ticketaka/mtvs3_final_backend/mail/query/repository/MailQueryRepository.java
@@ -1,4 +1,4 @@
-package ticketaka.mtvs3_final_backend.mail.command.domain.repository;
+package ticketaka.mtvs3_final_backend.mail.query.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,5 +7,7 @@ import ticketaka.mtvs3_final_backend.mail.command.domain.model.Mail;
 import java.util.List;
 
 @Repository
-public interface MailCommandRepository extends JpaRepository<Mail, Long> {
+public interface MailQueryRepository extends JpaRepository<Mail, Long> {
+
+    List<Mail> findAllByMemberId(Long memberId);
 }
